Add create() method for creating a new telemetry and element in their respective view configuration.
