+++ src/plugins/displayLayout/ElementViewConfiguration.js
@@ -25,6 +25,80 @@ define(
     function (ViewConfiguration) {
         class ElementViewConfiguration extends ViewConfiguration {
 
+            static create(type, openmct) {
+                const DEFAULT_DIMENSIONS = [2, 1],
+                      DEFAULT_X = 1,
+                      DEFAULT_Y = 1;
+                const INITIAL_STATES = {
+                    "image": {
+                        stroke: "transparent"
+                    },
+                    "box": {
+                        fill: "#717171",
+                        border: "transparent",
+                        stroke: "transparent"
+                    },
+                    "line": {
+                        x: 5,
+                        y: 3,
+                        x2: 6,
+                        y2: 6,
+                        stroke: "#717171"
+                    },
+                    "text": {
+                        fill: "transparent",
+                        stroke: "transparent",
+                        size: "13px"
+                    }
+                };
+                const DIALOGS = {
+                    "image": {
+                        name: "Image Properties",
+                        sections: [
+                            {
+                                rows: [
+                                    {
+                                        key: "url",
+                                        control: "textfield",
+                                        name: "Image URL",
+                                        "cssClass": "l-input-lg",
+                                        required: true
+                                    }
+                                ]
+                            }
+                        ]
+                    },
+                    "text": {
+                        name: "Text Element Properties",
+                        sections: [
+                            {
+                                rows: [
+                                    {
+                                        key: "text",
+                                        control: "textfield",
+                                        name: "Text",
+                                        required: true
+                                    }
+                                ]
+                            }
+                        ]
+                    }
+                };
+
+                let element = INITIAL_STATES[type] || {};
+                element = JSON.parse(JSON.stringify(element));
+                element.position = [
+                    element.x || DEFAULT_X,
+                    element.y || DEFAULT_Y
+                ];
+                element.dimensions = DEFAULT_DIMENSIONS;
+                element.type = type;
+
+                return DIALOGS[type] ?
+                    openmct.$injector.get('dialogService').getUserInput(DIALOGS[type], element) :
+                    element;
+            }
+
             /**
              * @param {Object} configuration the element (line, box, text or image) view configuration
              * @param {Object} configuration.element
@@ -33,6 +107,10 @@ define(
              * @param {Object} configuration.openmct the openmct object
              */
             constructor({element, ...rest}) {
+                rest.rawPosition = {
+                    position: element.position,
+                    dimensions: element.dimensions
+                };
                 super(rest);
                 this.element = element;
             }
