+++ src/plugins/displayLayout/components/LineView.vue
@@ -21,18 +21,18 @@
  *****************************************************************************/
 
  <template>
-    <!-- <svg width="{{ gridSize[0] * element.dimensions[0] }}"
-         height="{{ gridSize[1] * element.dimensions[1] }}">
-        <line x1="{{ gridSize[0] * element.x1 + 1 }}"
-              y1="{{ gridSize[1] * element.y1 + 1 }}"
-              x2="{{ gridSize[0] * element.x2 + 1 }}"
-              y2="{{ gridSize[1] * element.y2 + 1 }}"
-              stroke="{{ element.stroke }}"
+  <div>
+    <svg :width="gridSize[0] * element.dimensions[0]"
+         :height=" gridSize[1] * element.dimensions[1]">
+        <line :x1=" gridSize[0] * element.x1 + 1"
+              :y1="gridSize[1] * element.y1 + 1 "
+              :x2="gridSize[0] * element.x2 + 1"
+              :y2=" gridSize[1] * element.y2 + 1 "
+              :stroke="element.stroke"
               stroke-width="2">
         </line>
-    </svg> -->
-    <div>
-    </div>
+    </svg>
+  </div>
  </template>
 
  <script>
